third_party/WebKit/LayoutTests/http/tests/inspector/network/network-document-initiator.html

<html>
<head>
<link href="resources/initiator.css" rel="stylesheet" type="text/css">
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>
console.log("Page loaded.");

function navigateFromScript()
{
    window.location.href = "?foo";
}

var test = function()
{
    InspectorTest.addConsoleSniffer(step1);
    InspectorTest.evaluateInPage("navigateFromScript()");

    function findRequestByURL(url)
    {
        var requests = WebInspector.networkLog.requests;
        for (var i = 0; i < requests.length; ++i) {
            if (url.test(requests[i].url))
                return requests[i];
        }
    }

    function dumpInitiator(request)
    {
        var initiator = request.initiator;
        InspectorTest.addResult(request.url + ": " + initiator.type);
        if (initiator.url)
            InspectorTest.addResult("    " + initiator.url + " " + initiator.lineNumber);
        var stackTrace = initiator.stackTrace;
        if (!stackTrace)
            return;
        for (var i = 0; i < stackTrace.length; ++i) {
            var frame = stackTrace[i];
            if (frame.lineNumber) {
                InspectorTest.addResult("    " + frame.functionName + " " + frame.url + " " + frame.lineNumber);
                return;
            }
        }
    }

    function step1()
    {
        var request = findRequestByURL(/\?foo/);
        if (!request)
            return;
        dumpInitiator(request);
        InspectorTest.completeTest();
    }
}

</script>
</head>
<body onload="runTest()">
<p>Tests that page navigation initiated by JS is correctly reported.</p>
</body>
</html>
