third_party/WebKit/LayoutTests/http/tests/inspector/network/network-document-initiator-expected.txt

CONSOLE MESSAGE: line 7: Page loaded.
CONSOLE MESSAGE: line 7: Page loaded.
Tests that page navigation initiated by JS is correctly reported.

http://127.0.0.1:8000/inspector/network/network-document-initiator.html?foo: script
    navigateFromScript http://127.0.0.1:8000/inspector/network/network-document-initiator.html 11
